zebra: stop truncating the tc flower ethertype to 8 bits
netlink_tfilter_flower_request() stored htons(eth_proto) in a uint8_t
before passing it to nl_attr_put16(TCA_FLOWER_KEY_ETH_TYPE). The 16-bit
ethertype was truncated to one byte, so on little-endian hosts the high
byte of the (byte-swapped) value was kept and the low byte dropped -
e.g. ETH_P_IP (0x0800) became 0x08 and ETH_P_IPV6 (0x86dd) became 0x86.
The resulting TCA_FLOWER_KEY_ETH_TYPE attribute is garbage. Use a
uint16_t so the full ethertype is preserved.

+++ zebra/tc_netlink.c
@@ -364,7 +364,7 @@ static int netlink_tfilter_flower_put_options(struct nlmsghdr *n, size_t datalen
 {
 	struct inet_prefix addr;
 	uint32_t flags = 0, classid;
-	uint8_t protocol = htons(dplane_ctx_tc_filter_get_eth_proto(ctx));
+	uint16_t protocol = htons(dplane_ctx_tc_filter_get_eth_proto(ctx));
 	uint32_t filter_bm = dplane_ctx_tc_filter_get_filter_bm(ctx);
 
 	if (filter_bm & TC_FLOWER_SRC_IP) {
